+++ packages/label-core/src/rules/index.ts
@@ -36,3 +36,37 @@ export {
   GS1_DIGITAL_LINK_VALID,
   digitalLinkRule,
 } from './gs1/digitalLink'
+
+export {
+  GHS_LABEL_BELOW_MINIMUM_SIZE,
+  GHS_LABEL_SIZE_MET,
+  ghsLabelDimensionsRule,
+} from './ghs/labelDimensions'
+export {
+  GHS_PICTOGRAM_COMPLETE,
+  GHS_PICTOGRAM_NOT_RECOGNISED,
+  GHS_PICTOGRAM_SYMBOL_MISSING,
+  ghsPictogramIntegrityRule,
+} from './ghs/pictogramIntegrity'
+export {
+  GHS_PICTOGRAM_PRECEDENCE_MET,
+  GHS_PICTOGRAM_PRECEDENCE_OPTIONAL,
+  GHS_PICTOGRAM_PRECEDENCE_VIOLATED,
+  ghsPictogramPrecedenceRule,
+} from './ghs/pictogramPrecedence'
+export {
+  GHS_PICTOGRAM_BELOW_MINIMUM_SIZE,
+  GHS_PICTOGRAM_SIZE_MET,
+  ghsPictogramSizeRule,
+} from './ghs/pictogramSize'
+export {
+  GHS_SIGNAL_WORD_CONFLICT,
+  GHS_SIGNAL_WORD_SINGLE,
+  ghsSignalWordRule,
+} from './ghs/signalWord'
+export {
+  GHS_PICTOGRAM_MISSING,
+  GHS_PICTOGRAM_NOT_REQUIRED,
+  GHS_PICTOGRAM_SET_MATCHES,
+  ghsPictogramSetRule,
+} from './ghs/pictogramSet'